+++ crates/rust-client/src/store/web_store/mod.rs
@@ -180,14 +180,14 @@ impl Store for WebStore {
 
     async fn insert_account(
         &self,
-        account: &Account,
+        account: Account,
         account_seed: Option<Word>,
     ) -> Result<(), StoreError> {
         self.insert_account(account, account_seed).await
     }
 
-    async fn update_account(&self, new_account_state: &Account) -> Result<(), StoreError> {
-        self.update_account(new_account_state).await
+    async fn update_account(&self, new_account_state: Account) -> Result<(), StoreError> {
+        self.update_account(&new_account_state).await
     }
 
     async fn get_account_ids(&self) -> Result<Vec<AccountId>, StoreError> {