+++ CHANGELOG.md
@@ -67,6 +67,7 @@
 * [BREAKING] Removed `AccountId` to bech32 conversions and the `get_account_state_delta` RPC endpoint  ([#1177](https://github.com/0xMiden/miden-client/pull/1177)).
 * [BREAKING] Changed `exportNoteFile` to fail fast on invalid export type ([#1198](https://github.com/0xMiden/miden-client/pull/1198)).
 * [BREAKING] Refactored RPC errors ([#1202](https://github.com/0xMiden/miden-client/pull/1202)).
+* Avoid deep cloning of `Account` in  `Store`. ([1216](https://github.com/0xMiden/miden-client/pull/1216)).
 
 ## 0.10.2 (2025-08-04)